Move WithConsecutiveRector to its PHPUnit100 namespace
