+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/Fixture/avoid_nested_uses.php.inc
@@ -1,10 +1,10 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class AvoidUsingThisVariable extends TestCase
+final class AvoidNestedUses extends TestCase
 {
     public function test()
     {
@@ -14,7 +14,10 @@ final class AvoidUsingThisVariable extends TestCase
             ->method('prepare')
             ->withConsecutive(
                 [1, $value],
-                [3, $this->some(4)],
+                [2, function ($nested) {
+                    $nestedReturn = 1000;
+                    return $nestedReturn;
+                }],
             );
     }
 
@@ -28,11 +31,11 @@ final class AvoidUsingThisVariable extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class AvoidUsingThisVariable extends TestCase
+final class AvoidNestedUses extends TestCase
 {
     public function test()
     {
@@ -43,7 +46,10 @@ final class AvoidUsingThisVariable extends TestCase
             ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher, $value) {
             match ($matcher->numberOfInvocations()) {
                 1 => self::assertEquals([1, $value], $parameters),
-                2 => self::assertEquals([3, $this->some(4)], $parameters),
+                2 => self::assertEquals([2, function ($nested) {
+                    $nestedReturn = 1000;
+                    return $nestedReturn;
+                }], $parameters),
             };
         });
     }
