+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/Fixture/combine_with_on_consecutive_without_expects.php.inc
@@ -1,15 +1,14 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class CombineWithWillReturnOnConsecutiveCalls extends TestCase
+final class CombineWithOnConsecutiveWithoutExcepts extends TestCase
 {
     public function test()
     {
-        $this->userServiceMock->expects(self::exactly(2))
-            ->method('prepare')
+        $this->userServiceMock->method('prepare')
             ->withConsecutive(
                 [1, 2],
                 [3, 4],
@@ -22,17 +21,16 @@ final class CombineWithWillReturnOnConsecutiveCalls extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class CombineWithWillReturnOnConsecutiveCalls extends TestCase
+final class CombineWithOnConsecutiveWithoutExcepts extends TestCase
 {
     public function test()
     {
-        $matcher = self::exactly(2);
-        $this->userServiceMock->expects($matcher)
-            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher) {
+        $matcher = $this->exactly(2);
+        $this->userServiceMock->method('prepare')->willReturnCallback(function ($parameters) use ($matcher) {
             match ($matcher->numberOfInvocations()) {
                 1 => self::assertEquals([1, 2], $parameters),
                 2 => self::assertEquals([3, 4], $parameters),
