+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/Fixture/combine_with_will_return_self_this.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
@@ -8,8 +8,7 @@ final class CombineWithWillReturnSelfThis extends TestCase
 {
     public function test()
     {
-        $userServiceMock = $this->createMock(CombineWithWillReturnSelfThis::class);
-        $userServiceMock->expects(self::exactly(2))
+        $this->userServiceMock->expects(self::exactly(2))
             ->method('prepare')
             ->withConsecutive(
                 [1, 2],
@@ -23,7 +22,7 @@ final class CombineWithWillReturnSelfThis extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
@@ -31,15 +30,14 @@ final class CombineWithWillReturnSelfThis extends TestCase
 {
     public function test()
     {
-        $userServiceMock = $this->createMock(CombineWithWillReturnSelfThis::class);
         $matcher = self::exactly(2);
-        $userServiceMock->expects($matcher)
-            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher, $userServiceMock) {
+        $this->userServiceMock->expects($matcher)
+            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher) {
             match ($matcher->numberOfInvocations()) {
                 1 => self::assertEquals([1, 2], $parameters),
                 2 => self::assertEquals([3, 4], $parameters),
             };
-            return $userServiceMock;
+            return $this->userServiceMock;
         });
     }
 }
