+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/Fixture/include_use_of_variable.php.inc
@@ -1,17 +1,19 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class SomeTest extends TestCase
+final class IncludeUseOfVariable extends TestCase
 {
     public function test()
     {
+        $value = 1000;
+
         $this->personServiceMock->expects($this->exactly(2))
             ->method('prepare')
             ->withConsecutive(
-                [1, 2],
+                [1, $value],
                 [3, 4],
             );
     }
@@ -21,19 +23,21 @@ final class SomeTest extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class SomeTest extends TestCase
+final class IncludeUseOfVariable extends TestCase
 {
     public function test()
     {
+        $value = 1000;
         $matcher = $this->exactly(2);
+
         $this->personServiceMock->expects($matcher)
-            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher) {
+            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher, $value) {
             match ($matcher->numberOfInvocations()) {
-                1 => self::assertEquals([1, 2], $parameters),
+                1 => self::assertEquals([1, $value], $parameters),
                 2 => self::assertEquals([3, 4], $parameters),
             };
         });
