+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/Fixture/skip_will.php.inc
@@ -1,22 +1,21 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class SkipWillReturnMap extends TestCase
+final class SkipWill extends TestCase
 {
     public function test()
     {
+        $value = 5;
         $this->userServiceMock->expects(self::exactly(2))
             ->method('prepare')
             ->withConsecutive(
                 [1, 2],
                 [3, 4],
             )
-            ->willReturnMap([
-                [1, 2, 5],
-                [3, 4, 6],
-            ]);
+            ->will($this->returnValue($value));
     }
 }
+
