+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/Fixture/some_class.php.inc
@@ -1,19 +1,17 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class IncludeUseOfVariable extends TestCase
+final class SomeTest extends TestCase
 {
     public function test()
     {
-        $value = 1000;
-
         $this->personServiceMock->expects($this->exactly(2))
             ->method('prepare')
             ->withConsecutive(
-                [1, $value],
+                [1, 2],
                 [3, 4],
             );
     }
@@ -23,21 +21,19 @@ final class IncludeUseOfVariable extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
-final class IncludeUseOfVariable extends TestCase
+final class SomeTest extends TestCase
 {
     public function test()
     {
-        $value = 1000;
         $matcher = $this->exactly(2);
-
         $this->personServiceMock->expects($matcher)
-            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher, $value) {
+            ->method('prepare')->willReturnCallback(function ($parameters) use ($matcher) {
             match ($matcher->numberOfInvocations()) {
-                1 => self::assertEquals([1, $value], $parameters),
+                1 => self::assertEquals([1, 2], $parameters),
                 2 => self::assertEquals([3, 4], $parameters),
             };
         });
