+++ rules-tests/PHPUnit100/Rector/StmtsAwareInterface/WithConsecutiveRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\Rector\StmtsAwareInterface\WithConsecutiveRector;
+use Rector\PHPUnit\PHPUnit100\Rector\StmtsAwareInterface\WithConsecutiveRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(WithConsecutiveRector::class);
